[Issue #10627] (1/3) Refactor API key logic so it can be moved to grants_shared
