+++ api/src/auth/login_gov_jwt_auth.py
@@ -11,7 +11,7 @@ from grants_shared.auth.login_gov_jwt_auth import (
     get_config,
 )
 
-from src.auth.auth_handler import get_auth_handler
+from src.auth.auth_handler import AuthHandler
 
 logger = logging.getLogger(__name__)
 
@@ -52,6 +52,6 @@ def get_login_gov_redirect_uri(
     encoded_params = urllib.parse.urlencode(url_params)
 
     # Add the state to the DB
-    get_auth_handler(db_session).create_login_gov_state(state, nonce)
+    AuthHandler(db_session).create_login_gov_state(state, nonce)
 
     return f"{config.login_gov_auth_endpoint}?{encoded_params}"
