+++ api/src/services/users/rename_api_key.py
@@ -2,9 +2,9 @@ import logging
 from uuid import UUID
 
 from grants_shared.adapters import db
-from grants_shared.db.models.auth_base_models import BaseUserApiKey
 
-from src.services.users.get_user_api_keys import get_user_api_key
+from src.auth.api_key_handler import SimplerApiKeyHandler
+from src.db.models.user_models import UserApiKey
 
 logger = logging.getLogger(__name__)
 
@@ -18,17 +18,7 @@ class RenameApiKeyParams:
 
 def rename_api_key(
     db_session: db.Session, user_id: UUID, api_key_id: UUID, json_data: dict
-) -> BaseUserApiKey:
+) -> UserApiKey:
     """Rename an existing API key for a user"""
     params = RenameApiKeyParams(json_data)
-
-    api_key = get_user_api_key(db_session, user_id, api_key_id)
-
-    api_key.key_name = params.key_name
-
-    logger.info(
-        "Renamed API key",
-        extra=api_key.get_log_extra(),
-    )
-
-    return api_key
+    return SimplerApiKeyHandler(db_session).rename_api_key(user_id, api_key_id, params.key_name)
